从网络加载歌单

+++ app/src/main/java/com/sayi/vdim/sayi_music_entity/MusicFully.java
@@ -17,7 +17,7 @@ public class MusicFully {
         @SerializedName("artist")
         private String artist;
         @SerializedName("picid")
-        private int picId;
+        private long picId;
         @SerializedName("url")
         private String url;
         @SerializedName("pic")
